updated to render to glfw window

+++ src/bess/application/window.cpp
@@ -7,7 +7,6 @@
 #include "ext/vector_float2.hpp"
 #include "stb_image.h"
 #include "sub_systems/input_sub_system.h"
-#include "vulkan_core.h"
 #include <GLFW/glfw3.h>
 #include <cassert>
 #include <cstdint>
@@ -26,20 +25,7 @@ namespace Bess {
 
     void Window::onPreUpdate() { pollEvents(); }
 
-    void Window::onPreInit() {
-        initGLFW();
-
-        // auto vulkanCore =
-        //     GAppContext::getInstance().getSubSystem<Vulkan::VulkanCore>();
-        // vulkanCore->setWinExt(getVulkanExtensions());
-        //
-        // auto createSurface = [this](VkInstance &instance,
-        //                             VkSurfaceKHR &surface) {
-        //     createWindowSurface(instance, surface);
-        // };
-        //
-        // vulkanCore->setCreateSurfaceFn(createSurface);
-    }
+    void Window::onPreInit() { initGLFW(); }
 
     void Window::initGLFW() const {
         if (isGLFWInitialized)
@@ -125,7 +111,7 @@ namespace Bess {
                 const auto this_ = (Window *)glfwGetWindowUserPointer(window);
                 this_->m_framebufferResized = true;
 
-                Events::WindowResizeEvent evt{w, h};
+                Events::WindowResizeEvent evt{(uint32_t)w, (uint32_t)h};
 
                 auto &ctx = GAppContext::getInstance();
                 auto eventDispatcher =
